Keep HTLC balance in special address.
When tokens are locked into an HTLC, explicitly transfer them to
a special address (the contract address) instead of just deducting
them and later restoring them.

This makes sure that HTLC transfers correctly fit into the ERC-20
spec, while still the total supply is always correct and matches
up with the balances of all addresses.