Store only hashes of active HTLCs.
Instead of storing full data for all active HTLCs in the contract,
we only store a set (i.e. mapping to bool) of hashes of HTLC data.

This way, the users need to provide the data when timing out or
redeeming, but the contract can verify it is indeed matching an
active HTLC.  This saves contract storage and thus gas.