Add increase/decreaseAllowance methods.
Standard ERC-20 approval is vulnerable to a race-condition, where an
attacker can use up budget when a new approve transaction is being sent,
and the owner cannot reliably increase or decrease the allowance atomically
using just "approve".

This adds two new methods to the WCHI contract, which allow to
atomically increase or decrease the allowance by a given amount, independent
of the current allowance.  With these in conjunction with the normal
"approve" method, token owners can properly control the allowance atomically
for all relevant situations.